Change fsync(2) optimization to per-file
The zfs_fsyncer_key TSD was added as a performance optimization
to reduce contention on the zl_lock from zil_commit().  This
issue manifested itself as very long (100+ms) system call
times for fsync(2) heavy workloads.

However, we'd like to eliminate the usage of TSD in ZoL while
keeping this optimization.  There is no native Linux kernel
equivilant of TSD and the SPLs implementation is primitive.
It would simply be best to remove all the TSD consumers.

To achieve this the zfs_fsyncer_key TSD counter has been
reworked to be per-file instead of per-thread.  This is not
expected to behave significantly differently than the original
implementation for real workloads.

This just leaves one small ZFS TSD consumer.  If it can be
cleanly removed from the code we'll be able to shed the SPL
TSD implementation entirely.

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue zfsonlinux/spl#174