Do not use KERNEL_DIR environment variable in Makefile.am
A Gentoo user reported an issue where the build system would attempt to
recurse into the kernel source tree if KERNEL_DIR is set in the
environment. KERNEL_DIR is an environment variable that is used when the
kernel sources are in a non-standard location, so it is necessary to stop
relying on it to prevent this issue.

https://bugs.gentoo.org/show_bug.cgi?id=433946

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>