Revert "Fix using zvol as slog device"
This reverts commit 6c2856726fbae681649930d9620d9087080e58fc. It had the
effect of preventing deadlocks when ZFS was put on raw zvols, but the
deadlocks would still occur when something else such as LVM was between
the zvol and ZFS. In addition, it used lookup_bdev(). That broke support
for Debian Lenny's 2.6.26 kernel because it was not exported until Linux
2.6.27. That could be corrected with an autotools check, but we would
still have issues when something else is between the zvol and ZFS. As
such, it is best to revert this in favor of a different approach.

Signed-off-by: Richard Yao <ryao@gentoo.org>