Delete uxio loan code
NFS on Solaris and Illumos uses a ZFS-specific interface to avoid double
buffering. There is no hope of modifying NFS to use it on Linux, so it
serves no purpose to us. This code is also unused by other ZFS ports,
such as Mac ZFS and FreeBSD. Removing it makes the code that we do use
easier to maintain.

Signed-off-by: Richard Yao <ryao@cs.stonybrook.edu>